streamlit_app/requirements.txt
streamlit>=1.40.1
streamlit-option-menu>=0.4.0
altair>=5.4.1
abatar-0.2.2-py3-none-any.whl
pyspark
databricks-sqlalchemy
databricks-sdk
